requirements.txt
langgraph==0.2.45
langchain==0.3.7
langchain-google-vertexai==2.0.9
google-cloud-aiplatform==1.73.0
typing-extensions==4.12.2
pydantic==2.9.2
python-dotenv==1.0.1
